+++ doc/changelog.rst
@@ -15,6 +15,10 @@ Dropped support for Tornado 3. A recent version of Tornado 4 is required.
 
 Added support for `aiohttp`_ 2.0 and later, and dropped older aiohttp versions.
 
+New methods:
+
+- :meth:`MotorClient.list_databases` and :meth:`MotorClient.list_database_names`.
+
 The maximum number of workers in the thread pool can be overridden with an
 environment variable, see :doc:`configuration`.